Enhance Ethereum JSON-RPC API and Network Configuration
- Added new fields to BlockResponse for EIP-4844 and EIP-4895 support, including baseFeePerGas, withdrawalsRoot, and others.
- Implemented eth_getBlockReceipts, eth_feeHistory, eth_maxPriorityFeePerGas, and eth_blobBaseFee methods in EthBlocksService.
- Updated JsonRpcController to handle new JSON-RPC requests.
- Modified TransactionReceiptResponse and TransactionResponse to include new transaction fields for dynamic fees and blob transactions.
- Changed networkId type from Int to Long across various classes for better compatibility with larger network IDs.
- Improved error handling in Engine API server startup.
- Updated tests to reflect changes in networkId type and new API methods.

+++ src/main/scala/com/chipprbots/ethereum/network/NetworkPeerManagerActor.scala
@@ -539,7 +539,7 @@ object NetworkPeerManagerActor {
     */
   case class RemoteStatus(
       capability: Capability,
-      networkId: Int,
+      networkId: Long,
       chainWeight: ChainWeight,
       bestHash: ByteString,
       genesisHash: ByteString,
