+++ src/main/scala/com/chipprbots/ethereum/network/PeerManagerActor.scala
@@ -623,7 +623,7 @@ object PeerManagerActor {
     val waitForChainCheckTimeout: FiniteDuration
     val fastSyncHostConfiguration: FastSyncHostConfiguration
     val rlpxConfiguration: RLPxConfiguration
-    val networkId: Int
+    val networkId: Long
     val p2pVersion: Int
     val updateNodesInitialDelay: FiniteDuration
     val updateNodesInterval: FiniteDuration
