+++ src/main/scala/com/chipprbots/ethereum/utils/BlockchainConfig.scala
@@ -41,7 +41,7 @@ case class BlockchainConfig(
     daoForkConfig: Option[DaoForkConfig],
     accountStartNonce: UInt256,
     chainId: BigInt,
-    networkId: Int,
+    networkId: Long,
     monetaryPolicyConfig: MonetaryPolicyConfig,
     gasTieBreaker: Boolean,
     ethCompatibleStorage: Boolean,
@@ -175,9 +175,8 @@ object BlockchainConfig {
       parseHexOrDecNumber(s)
     }
 
-    val networkId: Int = Try(blockchainConfig.getInt("network-id")).getOrElse {
-      // Fallback for large network IDs (e.g., from hive test configs)
-      Try(BigInt(blockchainConfig.getString("network-id")).toInt).getOrElse(1)
+    val networkId: Long = Try(blockchainConfig.getLong("network-id")).getOrElse {
+      Try(BigInt(blockchainConfig.getString("network-id")).toLong).getOrElse(1L)
     }
 
     val monetaryPolicyConfig = MonetaryPolicyConfig(blockchainConfig.getConfig("monetary-policy"))
